refactor(test): P5-3 first cut — extract corner-radius tests from test_view.cpp
Phase 5 P5-3 first cut. Pull the 12-test pulp #1731 corner-radius
cluster (385 lines) out of test/test_view.cpp into a sibling
test_view_corner_radius.cpp TU.

Covers Panel + View percent-radius math, per-corner radius
resolution, reflow-on-bounds-change, view border stroke + outline +
box-shadow honoring effective_corner_radius.

- test/test_view.cpp: 2,170 → 1,778 lines (-392 net).
- test/test_view_corner_radius.cpp: 411 lines (new).
- New CMake target pulp-test-view-corner-radius via
  pulp_add_test_suite() (R2-7 helper).

Validation:
- pulp-test-view: 61 cases / 262 assertions — pass.
- pulp-test-view-corner-radius: 12 cases / 38 assertions — pass.

P5-3 progress: corner-radius slice done. Remaining slices (mask
fixtures, hit-test viewport extensions, layout cluster) can ship in
follow-up P5-3 PRs.
